+++ mobile/android/geckoview/src/main/java/org/mozilla/geckoview/doc-files/CHANGELOG.md
@@ -20,12 +20,15 @@ exclude: true
   `onTelemetryReceived` has been renamed to [`onHistogram`][71.4], and
   [`Metric`][71.5] now takes a type parameter.
   ([bug 1576730]({{bugzilla}}1576730))
-- Added overloads of [`GeckoSession.loadUri()`][71.6] that accept a map of 
+- Added overloads of [`GeckoSession.loadUri()`][71.6] that accept a map of
   additional HTTP request headers.
 - Added support for exposing the content blocking log in [`ContentBlockingController`][71.7].({{bugzilla}}1580201)
 - ⚠️  Added `nativeApp` to [`WebExtension.MessageDelegate.onMessage`][71.8] which
   exposes the native application identifier that was used to send the message.
   ({{bugzilla}}1546445)
+- Added [`GeckoRuntime.ServiceWorkerDelegate`][71.9] set via
+  ['setServiceWorkerDelegate'][71.10] to support ServiceWorkerClients.openWindow()
+  ([bug 1511033]({{bugzilla}}1511033))
 
 [71.1]: {{javadoc_uri}}/RuntimeTelemetry.Delegate.html#onBooleanScalar-org.mozilla.geckoview.RuntimeTelemetry.Metric-
 [71.2]: {{javadoc_uri}}/RuntimeTelemetry.Delegate.html#onLongScalar-org.mozilla.geckoview.RuntimeTelemetry.Metric-
@@ -35,6 +38,8 @@ exclude: true
 [71.6]: {{javadoc_uri}}/GeckoSession.html#loadUri-java.lang.String-java.io.File-java.util.Map-
 [71.7]: {{javadoc_uri}}/ContentBlockingController.html
 [71.8]: {{javadoc_uri}}/WebExtension.MessageDelegate.html#onMessage-java.lang.String-java.lang.Object-org.mozilla.geckoview.WebExtension.MessageSender-
+[71.9]: {{javadoc_uri}}/GeckoRuntime.ServiceWorkerDelegate.html
+[71.10]: {{javadoc_uri}}/GeckoRuntime#setServiceWorkerDelegate-org.mozilla.geckoview.GeckoRuntime.ServiceWorkerDelegate-
 
 ## v70
 - Added API for session context assignment
@@ -357,4 +362,4 @@ exclude: true
 [65.24]: {{javadoc_uri}}/CrashReporter.html#sendCrashReport-android.content.Context-android.os.Bundle-java.lang.String-
 [65.25]: {{javadoc_uri}}/GeckoResult.html
 
-[api-version]: 9fd93f90a9043b290ca9a3387228f00adfcedb51
+[api-version]: ec6b45461c040d8dc76aa0fed4917b3c4c07d67e